kyc-anti-injection-defence/AegisKYC-main/backend/app/utils/db.py
"""
Database connection utility module
"""
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

# MongoDB connection
MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://localhost:27017/")
client = MongoClient(MONGODB_URI)
db = client["aegis_kyc"]
