kyc-anti-injection-defence/AegisKYC-main/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
cryptography==41.0.7
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
paddleocr==2.7.0.3
pyzbar==0.1.9
python-Levenshtein==0.23.0
pytesseract==0.3.11
